pyproject.toml
[tool.poetry]
name = "swarm"
version = "0.1.0"
description = ""
authors = ["khizar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.39"
python-dotenv = "^1.0.1"
langchain-groq = "^0.2.0"
gradio = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
